feat: 完成事件处理架构重构
- 重构事件处理系统，移除复杂的Message层和MessageCategory层
- 实现统一消息类型架构，简化消息分发逻辑
- 新增provider_handler.rs和ui_handler.rs处理器
- 优化状态管理，减少代码复杂度
- 创建事件处理架构重构需求调研文档
- 提供详细的方案一实施方案文档
- 更新changelog记录重构进展

改进:
- 代码可维护性显著提升
- 消息处理逻辑更加清晰
- 减少了处理器间的耦合
- 提高了开发效率

+++ crates/domain_manager/src/gui/model/gui.rs
@@ -14,7 +14,7 @@ pub struct ReloadModel {
     pub domains: Vec<Domain>,
     pub records: Vec<DnsRecord>,
     pub message: String,
-    pub total_count: u64,
+    pub total_count: usize,
 }
 
 impl ReloadModel {
@@ -22,7 +22,7 @@ impl ReloadModel {
         providers: Vec<DomainProvider>,
         domains: Vec<Domain>,
         dns_records: Vec<DnsRecord>,
-        total_count: u64,
+        total_count: usize,
     ) -> Self {
         ReloadModel {
             reload_types: vec![],
